Changed CPU time back to Total time
I really didn't want to try to explain what this meant, and CPU time
wasn't really accurate either. It's measuring the amount of time a
specific database operation is taking, which includes the time waiting
for the database to respond. CPU time doesn't make it sound like it
includes that.

So I just changed the sentence to refer to the wall time measurements.